requirements.txt
streamlit>=1.30
google-generativeai>=0.6.0
python-dotenv>=1.0.0
requests>=2.31.0
